rgw: rest_client: work towards throttling of http read requests
Adjust the interfaces to provide the ability for the read callback
to pause the reads. While doing that, define a new class interface
for this instead of RGWGetDataCB. This had a butterfly effect that
required modifications to the obj read filters, but the end result
is a bit cleaner.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>
(cherry picked from commit b2143cded0e971361cdb089db19a6f69ce5b74dd)