rgw: remove warning
No need to clean io_ids when stack finishes, as list of io_ids is kept inside
the stack object, and will be cleaned up anyway. We also have a list of finished
ios in the completion manager (complete_reqs), but this is a fifo, and will be
cleaned up once pulled out (and we identify that io references an obsolete
stack).

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>
(cherry picked from commit b2d16317ba30b4d59662b79d01ac8f58697b7608)